重构

+++ src/main/java/org/confluence/mod/event/ItemEvents.java
@@ -165,7 +165,7 @@ public final class ItemEvents {
         ItemStack itemStack = event.getItem();
         if (itemStack.is(Tags.Items.TOOLS_BOWS) || itemStack.is(Tags.Items.TOOLS_CROSSBOWS)) {
             CompoundTag tag = itemStack.getTagElement(PrefixProvider.KEY);
-            if (tag != null && event.getEntity().level().getGameTime() % (int) (1.0 / tag.getDouble("attackSpeed")) == 0) {
+            if (tag != null && event.getEntity().level().getGameTime() % (int) (1.0 / tag.getFloat("attackSpeed")) == 0) {
                 event.setDuration(event.getDuration() - 1);
             }
         }
